Added DropDownListItemChargerWithScooterTest.java
